# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TurnByTurn API",
    "version": "1.0.0",
    "description": "Backend for TurnByTurn, a rotating savings (ajo / esusu) app. All money values are integer kobo. Errors always return `{ error: { code, message, details?, retryAfter? } }`; see docs/ERRORS.md for the code → screen map."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "NOT_FOUND",
                  "CONFLICT",
                  "INTERNAL_ERROR",
                  "GATEWAY_ERROR",
                  "INVALID_CREDENTIALS",
                  "OTP_INVALID",
                  "OTP_EXPIRED",
                  "OTP_RESEND_COOLDOWN",
                  "PASSWORD_POLICY",
                  "ACCOUNT_EXISTS",
                  "TOO_MANY_ATTEMPTS",
                  "BANK_VERIFICATION_FAILED",
                  "SESSION_EXPIRED",
                  "UNAUTHENTICATED",
                  "ACCESS_DENIED",
                  "SIGNUP_INCOMPLETE",
                  "GROUP_NOT_DRAFT",
                  "GROUP_NOT_ACTIVE",
                  "GROUP_SETUP_INCOMPLETE",
                  "GROUP_FULL",
                  "INVITE_INVALID",
                  "ALREADY_MEMBER",
                  "AMOUNT_INVALID",
                  "NOTHING_DUE",
                  "CYCLE_CLOSED",
                  "PAYOUT_NOT_ELIGIBLE",
                  "PAYOUT_ACCOUNT_MISSING",
                  "REMINDER_COOLDOWN"
                ]
              },
              "message": {
                "type": "string"
              },
              "retryAfter": {
                "type": "integer",
                "description": "Seconds; also sent as Retry-After header"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "Payout": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-f0-9]{24}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "blocked",
              "eligible",
              "processing",
              "sent",
              "failed",
              "delayed_recovery"
            ]
          },
          "expectedAmount": {
            "type": "integer",
            "description": "Amount in kobo (₦1 = 100)"
          },
          "confirmedAmount": {
            "type": "integer",
            "description": "Amount in kobo (₦1 = 100)"
          },
          "outstandingAmount": {
            "type": "integer",
            "description": "Amount in kobo (₦1 = 100)"
          },
          "reference": {
            "type": "string",
            "example": "PAY-0926-0003"
          },
          "destination": {
            "type": "object",
            "properties": {
              "bankCode": {
                "type": "string"
              },
              "bankName": {
                "type": "string"
              },
              "accountNumber": {
                "type": "string"
              },
              "accountName": {
                "type": "string"
              }
            }
          },
          "attempts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "reference": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "startedAt": {
                  "type": "string"
                },
                "finishedAt": {
                  "type": "string"
                },
                "failureReason": {
                  "type": "string"
                }
              }
            }
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          },
          "failureReason": {
            "type": "string"
          },
          "group": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "pattern": "^[a-f0-9]{24}$"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "cycle": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "pattern": "^[a-f0-9]{24}$"
              },
              "number": {
                "type": "integer"
              },
              "periodLabel": {
                "type": "string"
              },
              "dueDate": {
                "type": "string"
              },
              "status": {
                "type": "string"
              }
            }
          },
          "recipient": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "pattern": "^[a-f0-9]{24}$"
              },
              "name": {
                "type": "string"
              },
              "initials": {
                "type": "string"
              }
            }
          },
          "blockingMembers": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "membershipId": {
                  "type": "string",
                  "pattern": "^[a-f0-9]{24}$"
                },
                "name": {
                  "type": "string"
                },
                "outstanding": {
                  "type": "integer",
                  "description": "Amount in kobo (₦1 = 100)"
                }
              }
            }
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-f0-9]{24}$"
          },
          "type": {
            "type": "string",
            "enum": [
              "contribution",
              "fee",
              "payout",
              "refund",
              "adjustment"
            ]
          },
          "direction": {
            "type": "string",
            "enum": [
              "debit",
              "credit"
            ]
          },
          "amount": {
            "type": "integer",
            "description": "Amount in kobo (₦1 = 100)"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "success",
              "failed",
              "under_review",
              "reversed"
            ]
          },
          "reference": {
            "type": "string"
          },
          "groupId": {
            "type": "string",
            "pattern": "^[a-f0-9]{24}$"
          },
          "groupName": {
            "type": "string"
          },
          "cycleId": {
            "type": "string",
            "pattern": "^[a-f0-9]{24}$"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "meta": {
            "type": "object"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/home": {
      "get": {
        "tags": [
          "Home"
        ],
        "summary": "Member Home / Returning Home: cross-group overview",
        "operationId": "getHome",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "summary": {
                      "type": "object",
                      "properties": {
                        "groupCount": {
                          "type": "integer"
                        },
                        "coordinatingCount": {
                          "type": "integer"
                        },
                        "totalOutstanding": {
                          "type": "integer",
                          "description": "Amount in kobo (₦1 = 100)"
                        }
                      }
                    },
                    "nextDue": {
                      "type": "object",
                      "nullable": true
                    },
                    "upcomingPayout": {
                      "$ref": "#/components/schemas/Payout"
                    },
                    "groups": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "unreadNotifications": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED, SESSION_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Universal payment list grouped by month",
        "operationId": "getTransactions",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "2026-09"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "success",
                "failed",
                "under_review",
                "reversed"
              ]
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "contribution",
                "fee",
                "payout",
                "refund",
                "adjustment"
              ]
            }
          },
          {
            "name": "groupId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[a-f0-9]{24}$"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "months": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "month": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "totalOut": {
                            "type": "integer",
                            "description": "Amount in kobo (₦1 = 100)"
                          },
                          "totalIn": {
                            "type": "integer",
                            "description": "Amount in kobo (₦1 = 100)"
                          },
                          "transactions": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/Transaction"
                            }
                          }
                        }
                      }
                    },
                    "count": {
                      "type": "integer"
                    },
                    "nextBefore": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED, SESSION_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/filters": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Filter sheet values (period, status, type, group)",
        "operationId": "getTransactionsFilters",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "periods": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "month": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "statuses": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "types": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "groups": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "pattern": "^[a-f0-9]{24}$"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED, SESSION_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{id}": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Transaction details (+ receipt / payout)",
        "operationId": "getTransactionsId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transaction": {
                      "$ref": "#/components/schemas/Transaction"
                    },
                    "related": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      }
                    },
                    "receipt": {
                      "$ref": "#/components/schemas/Receipt"
                    },
                    "payout": {
                      "$ref": "#/components/schemas/Payout"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED, SESSION_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List notifications",
        "operationId": "getNotifications",
        "parameters": [
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notifications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notification"
                      }
                    },
                    "unreadCount": {
                      "type": "integer"
                    },
                    "nextBefore": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED, SESSION_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark read (ids or all)",
        "operationId": "postNotificationsRead",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updated": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED, SESSION_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^[a-f0-9]{24}$"
                    }
                  },
                  "all": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/support/tickets": {
      "post": {
        "tags": [
          "Support"
        ],
        "summary": "Help & Support: open a ticket",
        "operationId": "postSupportTickets",
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ticket": {
                      "$ref": "#/components/schemas/Ticket"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED, SESSION_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "general",
                      "payout",
                      "cycle_resolution"
                    ]
                  },
                  "message": {
                    "type": "string"
                  },
                  "groupId": {
                    "type": "string",
                    "pattern": "^[a-f0-9]{24}$"
                  },
                  "paymentReference": {
                    "type": "string"
                  }
                },
                "required": [
                  "message"
                ]
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Support"
        ],
        "summary": "My tickets",
        "operationId": "getSupportTickets",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tickets": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Ticket"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED, SESSION_EXPIRED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}